+++ src/util/translation.h
@@ -6,7 +6,7 @@
 #define BITCOIN_UTIL_TRANSLATION_H
 
 #include <tinyformat.h>
-
+#include <functional>
 
 /**
  * Bilingual messages:
